Small code optimizations -> added dedicated folder for objects and separate  folder for services
