+++ src/main/webapp/src/app/_domain-objects/objects.ts
@@ -1,0 +1,10 @@
+export class ObjectInfo {
+  objectName: string;
+  objectType: ObjectType;
+}
+
+export enum ObjectType {
+  STAR = 'star',
+  PLANETOID = 'planetoid',
+  COMET = 'comet'
+}
