+++ src/main/webapp/src/app/_domain-objects/user.ts
