--- src/main/webapp/src/app/objects.ts
@@ -1,4 +1,0 @@
-export class ObjectInfo {
-  objectName: string;
-  objectType: string;
-}
